airan-energy/requirements.txt
# Core JAX dependencies
jax>=0.4.20
jaxlib>=0.4.20
flax>=0.8.0
optax>=0.1.9
dm-haiku>=0.0.10

# Deep Learning utilities
chex>=0.1.85
jaxtyping>=0.2.25
equinox>=0.11.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Dashboard
streamlit>=1.28.0
dash>=2.14.0
dash-bootstrap-components>=1.5.0

# ML utilities
scikit-learn>=1.3.0
tensorboard>=2.14.0

# Time series
statsmodels>=0.14.0

# Config and logging
pyyaml>=6.0
tqdm>=4.65.0
python-dotenv>=1.0.0

# Experiment tracking (optional)
wandb>=0.15.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
